PMM work, more porting from master
* Rework 32-bit paging setup with proper higher-half mapping
* Add physical memory window for accessing high physical memory
  from 32-bit mode
* Expand heap allocator with better alignment and boundary handling

+++ kernel/arch/x86/boot/bios/paging32.h
@@ -4,8 +4,8 @@
 #include <stddef.h>
 
 
-void map_page_32(size_t size, u32 vaddr, u32 paddr, u32 flags, bool is_kernel);
-void map_region_32(size_t size, u32 vaddr, u32 paddr, u32 flags, bool is_kernel);
+void map_page_32(size_t size, u32 vaddr, u64 paddr, u64 flags, bool is_kernel);
+void map_region_32(size_t size, u32 vaddr, u64 paddr, u64 flags, bool is_kernel);
 void identity_map_32(u32 top_address, u32 offset, bool is_kernel);
 
 void setup_paging_32(void);
